requirements.txt
Django
djangorestframework
djangorestframework_simplejwt
django_cors_headers
Pillow
setuptools
numpy
scipy
gunicorn
packaging
psycopg2-binary
pytz